# swagger/swagger.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: BookLibrary
  description: |
    books available for loan
  license:
    name: MIT
servers:
- url: http://localhost:3000
paths:
...
  /books/{bookId}:
    get:
      summary: get one book by id
      operationId: bookByID
      tags: 
        - books
      parameters: 
       - name: bookId
         in: path
         required: true
         description: The ID of the book
         schema:
           title: string
      responses:
        '200':
          description: You have fetched the book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseBook"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update one Book
      operationId: updateBook
      tags: 
        - books
      requestBody:
        description: update one book
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      parameters: 
      - name: bookId
        in: path
        required: true
        description: The ID of the book
        schema:
          type: string
      responses:
        '200':
          description: Book Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBook'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete A book
      operationId: deleteBook
      tags: 
        - books
      parameters: 
      - name: bookId
        in: path
        required: true
        description: Deletes the Book
        schema:
          type: string
      responses:
        '200':
          description: Book Deleted!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBook'
        '204':
          description: Nothing to delete!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Book: # A single book
      type: object
      properties:
# Add properties for your books here   
        isbn:
          type: string
        title:
          type: string
        author:
          type: string
        price:
          type: number
        selleremail:
         type: string
        used:
          type: boolean
        location:
          properties:
            city:
              type: string
            street:
              type: string
    ResponseBook: #Single book response including _id and __v
      allOf:
      - $ref: "#/components/schemas/Book"
      - properties:
          __v:
            type: integer
            format: uint32
          _id:
            type: string
            format: uuid
...
    Error:
      required:
        - error
      properties:
        error:
          type: object
          properties:
            expose:
              type: boolean
            statusCode:
              type: integer
              format: uint32
            status:
              type: integer
              format: uint32
            body:
              title: body as escaped json
              type: string
            type:
              title: fault desciption
              type: string